src/main/java/it/polimi/ingsw/model/Turn.java
package it.polimi.ingsw.model;

import java.util.List;

public class Turn
{
    private int turnNumber;
    private List<Worker> workers;
    private int validMoves;

    public int getTurn() {
        return turnNumber;
    }

    public void setTurn(int turn) {
        this.turnNumber = turn;
    }

    public int getValidMoves() {
        return validMoves;
    }

    public void setValidMoves(int validMoves) {
        this.validMoves = validMoves;
    }

    public List<Worker> getWorkers() {
        return workers;
    }

    public void setWorkers(List<Worker> workers) {
        this.workers = workers;
    }

    public void StartTurn()
    {

    }
    public void Move()
    {

    }
    public void Build()
    {

    }
    public void CheckLostOthers()
    {

    }
    public void CheckWinCondition()
    {

    }
}